+++ frontend/src/pages/DataCatalog/dataset_explorer/storybook/DataSourceListing.stories.tsx
@@ -6,7 +6,7 @@ import {
 } from "../DataSourceListing";
 import { StoryWrapper } from "../../../../storybook/StoryWrapper";
 import { MapOfDatasetMetadata } from "../../../../data/utils/DatasetTypes";
-import { FakeDatasetMetadataMap } from "../../../../data/config/MetadataMap";
+import { FakeDatasetMetadataMap } from "../../../../data/config/FakeDatasetMetadata";
 
 export default {
   title: "DatasetExplorer/DataSourceListing",
